Criado componentes e tela da ChatPage, além de inserido o sistema de listagem e envio de mensagens - SEM O SIGNALR

+++ ChatNow_WebAPi/Repositories/FriendshipRepository.cs
@@ -16,6 +16,26 @@ namespace ChatNow_WebAPi.Repositories
             _context = new();
         }
 
+        public Friendship AcceptFriendInvite(Guid idFriendship)
+        {
+            Friendship searchedFriendship = _context.Friendships.FirstOrDefault(f => f.IdFriendship == idFriendship)!;
+
+            if (searchedFriendship is null)
+                return null;
+
+            // Altera o id do status "Confirmado"
+            searchedFriendship.StatusId = Guid.Parse("4F122E07-7A16-47C4-BE2C-6671848C8417");
+
+            // Atualiza a Friendship
+            _context.Friendships.Update(searchedFriendship);
+
+            // Salva o banco de dados
+            _context.SaveChanges();
+
+            // retorna o objeto atualizado
+            return searchedFriendship;
+        }
+
 
         public async void AddFriend(Guid idActualUser, Guid idPendentUser)
         {
@@ -37,20 +57,23 @@ namespace ChatNow_WebAPi.Repositories
             if (!friendshipAlredyExists)
                 _context.Friendships.Add(newFriendship);
 
-            // Cria uma nova Conversa
-            Guid newConversationId = await CreateConversation();
+            // Cria uma nova Conversa com o Id da Friendship criada
+            Guid newConversationId = await CreateConversation(newFriendship.IdFriendship);
 
             // Cria uma referência do Usuário atual e do Remetente à Conversa 
-            CreateUserConversations([idActualUser, idPendentUser], newConversationId);
+            //CreateUserConversations([idActualUser, idPendentUser], newConversationId);
 
             // Salva as alterações
             _context.SaveChanges();
         }
 
-        private async Task<Guid> CreateConversation()
+        private async Task<Guid> CreateConversation(Guid idFriendship)
         {
             // Cria uma nova conversa
-            var newConversation = new Conversation();
+            var newConversation = new Conversation
+            {
+                IdFriendship = idFriendship
+            };
 
             // Adiciona a nova conversa ao contexto
             await _context.Conversations.AddAsync(newConversation);
@@ -60,31 +83,34 @@ namespace ChatNow_WebAPi.Repositories
 
             return newConversation.Id;
         }
-        private void CreateUserConversations(Guid[] userIds, Guid idConversation)
-        {
-            foreach (var userId in userIds)
-            {
-                var newUserConversation = new UserConversation()
-                {
-                    IdUser = userId,
-                    IdConversation = idConversation
-                };
+        //private void CreateUserConversations(Guid[] userIds, Guid idConversation)
+        //{
+        //    foreach (var userId in userIds)
+        //    {
+        //        var newUserConversation = new UserConversation()
+        //        {
+        //            IdUser = userId,
+        //            IdConversation = idConversation
+        //        };
 
-                _context.UserConversation.Add(newUserConversation);
-            }
-        }
+        //        _context.UserConversation.Add(newUserConversation);
+        //    }
+        //}
 
 
         public void RemoveFriend(Guid idFriendship)
         {
             var friendshipToDelete = _context.Friendships.FirstOrDefault(f => f.IdFriendship == idFriendship);
+            var conversationToDelete = _context.Conversations.FirstOrDefault(c => c.IdFriendship == idFriendship);
 
-            if (friendshipToDelete is not null)
+            if (friendshipToDelete is not null && conversationToDelete is not null)
             {
+
                 // Continuar lógica de remoção da Conversa
-                _context.UserConversation.FirstOrDefault(uc => uc.IdUser == friendshipToDelete.IdUserOne || uc.IdUser == friendshipToDelete.IdUserTwo)
+                //_context.UserConversation.FirstOrDefault(uc => uc.IdUser == friendshipToDelete.IdUserOne || uc.IdUser == friendshipToDelete.IdUserTwo);
 
                 _context.Friendships.Remove(friendshipToDelete);
+                _context.Conversations.Remove(conversationToDelete);
             }
 
             _context.SaveChanges();
